+++ cards/exports.ts
@@ -16,10 +16,10 @@ export { blueprint as c0270333 } from './Classes/Druid/Collectible/Spells/4-Cost
 export { blueprint as c1784065 } from './Classes/Druid/Collectible/Spells/4-Cost/87-oaken-summons.js';
 export { blueprint as cc8d1f87 } from './Classes/Druid/Collectible/Spells/5-Cost/97-flipper-friends.js';
 export { blueprint as cc90c788 } from './Classes/Druid/Collectible/Spells/7-Cost/100-scale-of-onyxia.js';
+export { blueprint as c80d149d } from './Classes/Druid/Uncollectible/Heropowers/2-Cost/113-nurture.js';
 export { blueprint as c30d97aa } from './Classes/Druid/Uncollectible/Minions/1-Cost/83-treant.js';
 export { blueprint as c7245d5c } from './Classes/Druid/Uncollectible/Minions/1-Cost/95-otter.js';
 export { blueprint as cbca7e23 } from './Classes/Druid/Uncollectible/Minions/6-Cost/96-orca.js';
-export { blueprint as c94212df } from './Classes/Druid/Uncollectible/Heropowers/113-nurture.js';
 export { blueprint as c9841922 } from './Classes/Mage/Collectible/Spells/10-Cost/105-pyroblast.js';
 export { blueprint as c3dee25b } from './Classes/Neutral/Collectible/Locations/7-Cost/102-prison-of-yogg-saron.js';
 export { blueprint as c28b7adc } from './Classes/Neutral/Collectible/Minions/1-Cost/110-chaotic-tendril.js';
